try to be clearer about volunteers and funding
We added the Core Team section to highlight people who have made
significant contributions to Elm, whether that was in code or in
emotional labor.

I think this may have given the impression that we are more organized
than we actually are. The only person currently paid to work full-time
is me, and others do what they can while prioritizing their family and
personal health. I think folks also thought that we had bandwidth to
manage more work or different work than is the real reality.

I hope that the new version helps people understand what is going on
better.