add Large Records section / remove extensiblity stuff
We used to think that making type alias of extensible records might be a cool idea, but did not have much usage experience to decide one way or the other. Enough people have tried it by now that It seems like it is not a great way to structure code. It is fine in some cases, but having explicit structures for each phase seems to come together nicer overall. **Better to have a function that takes four arguments than a function that takes an open record with four specific fields.** Just pass the four values in, then it doesn't matter whether the data structure is a record or custom type or whatever else.

Anyway, I replaced it with some notes on "Large Records" to try to warn people about the `Bool` and `Maybe` pitfalls. Thanks to @JoelQ for the nice posts describing these cases!